+++ nsxt/resource_nsxt_policy_transit_gateway_nat_rule.go
@@ -12,12 +12,14 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/vmware/vsphere-automation-sdk-go/runtime/protocol/client"
 	"github.com/vmware/vsphere-automation-sdk-go/services/nsxt/model"
-	clientLayer "github.com/vmware/vsphere-automation-sdk-go/services/nsxt/orgs/projects/transit_gateways/nat"
 
+	"github.com/vmware/terraform-provider-nsxt/api/orgs/projects/transit_gateways/nat"
 	utl "github.com/vmware/terraform-provider-nsxt/api/utl"
 	"github.com/vmware/terraform-provider-nsxt/nsxt/metadata"
 	"github.com/vmware/terraform-provider-nsxt/nsxt/util"
 )
+
+var cliTransitGatewayNatRulesClient = nat.NewNatRulesClient
 
 var transitGatewayNatPathExample = "/orgs/[org]/projects/[project]/transit-gateways/[gateway]/nat/[type]"
 
@@ -42,7 +44,10 @@ func resourceNsxtPolicyTransitGatewayNatRuleExists(sessionContext utl.SessionCon
 	if pathErr != nil {
 		return false, pathErr
 	}
-	client := clientLayer.NewNatRulesClient(connector)
+	client := cliTransitGatewayNatRulesClient(sessionContext, connector)
+	if client == nil {
+		return false, fmt.Errorf("unsupported client type")
+	}
 	_, err = client.Get(parents[0], parents[1], parents[2], parents[3], id)
 	if err == nil {
 		return true, nil
@@ -88,7 +93,11 @@ func resourceNsxtPolicyTransitGatewayNatRuleCreate(d *schema.ResourceData, m int
 
 	log.Printf("[INFO] Creating PolicyTransitGatewayNatRule with ID %s", id)
 
-	client := clientLayer.NewNatRulesClient(connector)
+	sessionContext := getParentContext(d, m, parentPath)
+	client := cliTransitGatewayNatRulesClient(sessionContext, connector)
+	if client == nil {
+		return fmt.Errorf("unsupported client type")
+	}
 	err = client.Patch(parents[0], parents[1], parents[2], parents[3], id, obj)
 	if err != nil {
 		return handleCreateError("PolicyTransitGatewayNatRule", id, err)
@@ -107,8 +116,12 @@ func resourceNsxtPolicyTransitGatewayNatRuleRead(d *schema.ResourceData, m inter
 		return fmt.Errorf("Error obtaining PolicyTransitGatewayNatRule ID")
 	}
 
-	client := clientLayer.NewNatRulesClient(connector)
 	parentPath := d.Get("parent_path").(string)
+	sessionContext := getParentContext(d, m, parentPath)
+	client := cliTransitGatewayNatRulesClient(sessionContext, connector)
+	if client == nil {
+		return fmt.Errorf("unsupported client type")
+	}
 	parents, pathErr := parseStandardPolicyPathVerifySize(parentPath, 4, transitGatewayNatPathExample)
 	if pathErr != nil {
 		return pathErr
@@ -160,7 +173,11 @@ func resourceNsxtPolicyTransitGatewayNatRuleUpdate(d *schema.ResourceData, m int
 	if err := metadata.SchemaToStruct(elem, d, getPolicyVpcNatRuleSchema(true), "", nil); err != nil {
 		return err
 	}
-	client := clientLayer.NewNatRulesClient(connector)
+	sessionContext := getParentContext(d, m, parentPath)
+	client := cliTransitGatewayNatRulesClient(sessionContext, connector)
+	if client == nil {
+		return fmt.Errorf("unsupported client type")
+	}
 	_, err := client.Update(parents[0], parents[1], parents[2], parents[3], id, obj)
 	if err != nil {
 		return handleUpdateError("PolicyTransitGatewayNatRule", id, err)
@@ -182,7 +199,11 @@ func resourceNsxtPolicyTransitGatewayNatRuleDelete(d *schema.ResourceData, m int
 		return pathErr
 	}
 
-	client := clientLayer.NewNatRulesClient(connector)
+	sessionContext := getParentContext(d, m, parentPath)
+	client := cliTransitGatewayNatRulesClient(sessionContext, connector)
+	if client == nil {
+		return fmt.Errorf("unsupported client type")
+	}
 	err := client.Delete(parents[0], parents[1], parents[2], parents[3], id)
 
 	if err != nil {
